Setup minimum bound to Go 1.18 (#110)
Set minimum Go version to 1.18 to enable pruning as described by #109.
1.18 is already ancient in its own respect, so the good procured by
making it the minimum far outweighs any compatibility problems.

Update the build matrix to use the latest Go 1.25 and next latest
versions of 1.24 and 1.23. Add 1.18 in there as well to make sure that
actually works.

I originally had this set to minimum of Go 1.17 because that's the
version where module pruning was introduced, but it turns out that all
kinds of dependencies are using functions like `strings.Cut` that were
introduced in 1.18. Do 1.18 instead to make this easier.

Fixes #109.